Move circuit REST API resources to admin rest_api module
This moves and splits up the circuit REST API resources and into
actix-specific modules.
