+++ libsplinter/src/admin/rest_api/actix/mod.rs
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#[cfg(feature = "circuit-read")]
+pub(super) mod circuits;
+#[cfg(feature = "circuit-read")]
+pub(super) mod circuits_circuit_id;
 #[cfg(feature = "proposal-read")]
 pub(super) mod proposals_read;
 #[cfg(feature = "proposal-read")]
