+++ libsplinter/src/admin/rest_api/mod.rs
@@ -19,8 +19,13 @@ mod error;
 mod resources;
 
 use crate::admin::service::AdminService;
+use crate::circuit::store;
 use crate::rest_api::{Resource, RestResourceProvider};
 
+#[cfg(all(feature = "circuit-read", feature = "rest-api-actix"))]
+use self::actix::circuits::make_list_circuits_resource;
+#[cfg(all(feature = "circuit-read", feature = "rest-api-actix"))]
+use self::actix::circuits_circuit_id::make_fetch_circuit_resource;
 #[cfg(all(feature = "proposal-read", feature = "rest-api-actix"))]
 use self::actix::proposals_read::make_list_proposals_resource;
 #[cfg(all(feature = "proposal-read", feature = "rest-api-actix"))]
@@ -44,6 +49,34 @@ impl RestResourceProvider for AdminService {
         #[cfg(all(feature = "proposal-read", feature = "rest-api-actix"))]
         resources.push(make_list_proposals_resource(self.commands()));
 
+        resources
+    }
+}
+
+#[derive(Clone)]
+pub struct CircuitResourceProvider<T: store::CircuitStore> {
+    node_id: String,
+    store: T,
+}
+
+impl<T: store::CircuitStore + 'static> CircuitResourceProvider<T> {
+    pub fn new(node_id: String, store: T) -> Self {
+        Self { node_id, store }
+    }
+}
+
+impl<T: store::CircuitStore + 'static> RestResourceProvider for CircuitResourceProvider<T> {
+    fn resources(&self) -> Vec<Resource> {
+        // Allowing unused_mut because resources must be mutable if feature circuit-read is enabled
+        #[allow(unused_mut)]
+        let mut resources = Vec::new();
+        #[cfg(feature = "circuit-read")]
+        {
+            resources.append(&mut vec![
+                make_fetch_circuit_resource(self.store.clone()),
+                make_list_circuits_resource(self.store.clone()),
+            ])
+        }
         resources
     }
 }
