+++ libsplinter/src/circuit/mod.rs
@@ -14,8 +14,6 @@
 
 pub mod directory;
 pub mod handlers;
-#[cfg(feature = "rest-api")]
-pub mod rest_api;
 pub mod service;
 pub mod store;
